+++ mne/cov.py
@@ -4,21 +4,18 @@
 #
 # License: BSD (3-clause)
 
-from contextlib import contextmanager
 from copy import deepcopy
 import itertools as itt
 from math import log
-import operator
 import os
 
 import numpy as np
 from scipy import linalg, sparse
 
-from .io.meas_info import _simplify_info
 from .io.write import start_file, end_file
 from .io.proj import (make_projector, _proj_equal, activate_proj,
                       _check_projs, _needs_eeg_average_ref_proj,
-                      _has_eeg_average_ref_proj)
+                      _has_eeg_average_ref_proj, _read_proj, _write_proj)
 from .io import fiff_open
 from .io.pick import (pick_types, pick_channels_cov, pick_channels, pick_info,
                       _picks_by_type, _pick_data_channels, _picks_to_idx,
@@ -26,7 +23,6 @@ from .io.pick import (pick_types, pick_channels_cov, pick_channels, pick_info,
 
 from .io.constants import FIFF
 from .io.meas_info import read_bad_channels, create_info
-from .io.proj import _read_proj, _write_proj
 from .io.tag import find_tag
 from .io.tree import dir_tree_find
 from .io.write import (start_block, end_block, write_int, write_name_list,
@@ -34,13 +30,11 @@ from .io.write import (start_block, end_block, write_int, write_name_list,
 from .defaults import _handle_default
 from .epochs import Epochs
 from .event import make_fixed_length_events
-from .utils import (check_fname, logger, verbose,
-                    check_version, _time_mask, warn,
-                    copy_function_doc_to_method_doc, _pl,
-                    _undo_scaling_cov, _undo_scaling_array,
-                    _apply_scaling_array, _check_rank_cov)
+from .rank import compute_rank
+from .utils import (check_fname, logger, verbose, check_version, _time_mask,
+                    warn, copy_function_doc_to_method_doc, _pl,
+                    _undo_scaling_cov, _scaled_array)
 from . import viz
-from .rank import _estimate_rank_meeg_cov
 
 from .fixes import BaseEstimator, EmpiricalCovariance, _logdet
 
@@ -396,16 +390,12 @@ def compute_raw_covariance(raw, tmin=0, tmax=None, tstep=0.2, reject=None,
         rejected. If False, no rejection based on annotations is performed.
 
         .. versionadded:: 0.14
-    rank : None | int | dict | 'full'
-        Specified rank of the noise covariance matrix. If None(default),
-        the rank is detected automatically.
-        If int, the rank is specified for the MEG
-        channels. A dictionary with entries 'eeg', 'meg' or any other
-        data channel type such as 'seeg' or 'ecog' can be used
-        to specify the rank for each modality. If 'full',
-        the covariance is assumed to be full-rank when regularizing.
+    %(rank_None)s
 
         .. versionadded:: 0.17
+
+        .. versionadded:: 0.18
+           Support for 'info' mode.
     %(verbose)s
 
     Returns
@@ -542,7 +532,13 @@ def _check_method_params(method, method_params, keep_sample_mean=True,
             'Invalid {name} ({method}). Accepted values (individually or '
             'in a list) are any of "{accepted_methods}" or None.'.format(
                 name=name, method=method, accepted_methods=accepted_methods))
-    rank, method = _check_rank_cov(rank, method, was_auto)
+    if not (isinstance(rank, str) and rank == 'full'):
+        if was_auto:
+            method.pop(method.index('factor_analysis'))
+        for method_ in method:
+            if method_ in ('pca', 'factor_analysis'):
+                raise ValueError('%s can so far only be used with rank="full",'
+                                 ' got rank=%r' % (method_, rank))
     if not keep_sample_mean:
         if len(method) != 1 or 'empirical' not in method:
             raise ValueError('`keep_sample_mean=False` is only supported'
@@ -551,7 +547,7 @@ def _check_method_params(method, method_params, keep_sample_mean=True,
             if v.get('assume_centered', None) is False:
                 raise ValueError('`assume_centered` must be True'
                                  ' if `keep_sample_mean` is False')
-    return method, _method_params, rank
+    return method, _method_params
 
 
 @verbose
@@ -641,16 +637,12 @@ def compute_covariance(epochs, keep_sample_mean=True, tmin=None, tmax=None,
         unstable results in covariance calculation, e.g. when data
         have been processed with Maxwell filtering but not transformed
         to the same head position.
-    rank : None | int | dict | 'full'
-        Specified rank of the noise covariance matrix. If None (default),
-        the rank is detected automatically.
-        If int, the rank is specified for the MEG
-        channels. A dictionary with entries 'eeg', 'meg' or any other
-        data channel type such as 'seeg' or 'ecog' can be used
-        to specify the rank for each modality. If 'full',
-        the covariance is assumed to be full-rank when regularizing.
+    %(rank_None)s
 
         .. versionadded:: 0.17
+
+        .. versionadded:: 0.18
+           Support for 'info' mode.
     %(verbose)s
 
     Returns
@@ -738,7 +730,7 @@ def compute_covariance(epochs, keep_sample_mean=True, tmin=None, tmax=None,
     """
     # scale to natural unit for best stability with MEG/EEG
     scalings = _check_scalings_user(scalings)
-    method, _method_params, rank = _check_method_params(
+    method, _method_params = _check_method_params(
         method, method_params, keep_sample_mean, rank=rank)
     del method_params
 
@@ -909,38 +901,25 @@ def _eigvec_subspace(eig, eigvec, mask):
     return eig, eigvec
 
 
-@contextmanager
-def _scaled_array(data, picks_list, scalings):
-    """Scale, use, unscale array."""
-    _apply_scaling_array(data.T, picks_list=picks_list, scalings=scalings)
-    try:
-        yield
-    finally:
-        _undo_scaling_array(data.T, picks_list=picks_list, scalings=scalings)
-
-
 def _compute_covariance_auto(data, method, info, method_params, cv,
                              scalings, n_jobs, stop_early, picks_list, rank):
     """Compute covariance auto mode."""
     # rescale to improve numerical stability
-    with _scaled_array(data, picks_list, scalings):
-        if rank != 'full':
-            C = np.dot(data.T, data)
-            # already scaled
-            _, eigvec, mask = _smart_eigh(C, info, rank, scalings=1.,
-                                          proj_subspace=True)
-            eigvec = eigvec[mask]
-            data = np.dot(data, eigvec.T)
-            used = np.where(mask)[0]
-            sub_picks_list = [(key, np.searchsorted(used, picks))
-                              for key, picks in picks_list]
-            sub_info = pick_info(info, used)
-            logger.info('Reducing data rank from %s -> %s'
-                        % (len(mask), eigvec.shape[0]))
-        else:
-            eigvec = None
-            sub_picks_list = picks_list
-            sub_info = info
+    from .io import RawArray
+    raw_temp = RawArray(data.T, info)
+    orig_rank = rank
+    rank = compute_rank(raw_temp, rank, scalings, info)
+    with _scaled_array(data.T, picks_list, scalings):
+        C = np.dot(data.T, data)
+        _, eigvec, mask = _smart_eigh(C, info, rank, proj_subspace=True)
+        eigvec = eigvec[mask]
+        data = np.dot(data, eigvec.T)
+        used = np.where(mask)[0]
+        sub_picks_list = [(key, np.searchsorted(used, picks))
+                          for key, picks in picks_list]
+        sub_info = pick_info(info, used)
+        logger.info('Reducing data rank from %s -> %s'
+                    % (len(mask), eigvec.shape[0]))
         estimator_cov_info = list()
         msg = 'Estimating covariance using %s'
 
@@ -1022,7 +1001,7 @@ def _compute_covariance_auto(data, method, info, method_params, cv,
                 del shrinkage, sc
 
             elif method_ == 'pca':
-                assert rank == 'full'  # guaranteed above
+                assert orig_rank == 'full'
                 pca, _info = _auto_low_rank_model(
                     data_, method_, n_jobs=n_jobs, method_params=mp, cv=cv,
                     stop_early=stop_early)
@@ -1031,7 +1010,7 @@ def _compute_covariance_auto(data, method, info, method_params, cv,
                 del pca
 
             elif method_ == 'factor_analysis':
-                assert rank == 'full'
+                assert orig_rank == 'full'
                 fa, _info = _auto_low_rank_model(
                     data_, method_, n_jobs=n_jobs, method_params=mp, cv=cv,
                     stop_early=stop_early)
@@ -1053,9 +1032,8 @@ def _compute_covariance_auto(data, method, info, method_params, cv,
                 loglik = _cross_val(data, estimator, cv, n_jobs)
             else:
                 loglik = None
-            if eigvec is not None:
-                # project back if necessary
-                cov = np.dot(eigvec.T, np.dot(cov, eigvec))
+            # project back
+            cov = np.dot(eigvec.T, np.dot(cov, eigvec))
             # undo scaling
             _undo_scaling_cov(cov, picks_list, scalings)
             method_ = method[ei]
@@ -1331,8 +1309,8 @@ def _get_whitener(noise_cov, info=None, ch_names=None, rank=None,
     #   Handle noise cov
     #
     if not prepared:
-        noise_cov = prepare_noise_cov(noise_cov, info,
-                                      ch_names, rank, scalings)
+        noise_cov = prepare_noise_cov(noise_cov, info, ch_names, rank,
+                                      scalings)
     n_chan = len(noise_cov['eig'])
 
     #   Omit the zeroes due to projection
@@ -1369,12 +1347,10 @@ def prepare_noise_cov(noise_cov, info, ch_names=None, rank=None,
     ch_names : list | None
         The channel names to be considered. Can be None to use
         ``info['ch_names']``.
-    rank : None | int | dict (default None)
-        Specified rank of the noise covariance matrix. If None, the rank is
-        detected automatically. If int, the rank is specified for the MEG
-        channels. A dictionary with entries 'eeg', 'meg' or any other
-        data channel type such as 'seeg' or 'ecog' can be used
-        to specify the rank for each modality.
+    %(rank_None)s
+
+        .. versionadded:: 0.18
+           Support for 'info' mode.
     scalings : dict | None
         Data will be rescaled before rank estimation to improve accuracy.
         If dict, it will override the following dict (default if None)::
@@ -1396,22 +1372,25 @@ def prepare_noise_cov(noise_cov, info, ch_names=None, rank=None,
         C = noise_cov.data[np.ix_(noise_cov_idx, noise_cov_idx)]
     else:
         C = np.diag(noise_cov.data[noise_cov_idx])
+    info = pick_info(info, pick_channels(info['ch_names'], ch_names))
     projs = info['projs'] + noise_cov['projs']
-    scalings = _handle_default('scalings_cov_rank', scalings)
-    eig, eigvec, _ = _smart_eigh(C, info, rank, scalings, projs, ch_names)
-
     noise_cov = Covariance(
         data=C, names=ch_names, bads=list(noise_cov['bads']),
-        projs=deepcopy(noise_cov['projs']),
-        nfree=noise_cov['nfree'], eig=eig, eigvec=eigvec,
+        projs=deepcopy(noise_cov['projs']), nfree=noise_cov['nfree'],
         method=noise_cov.get('method', None),
         loglik=noise_cov.get('loglik', None))
+
+    eig, eigvec, _ = _smart_eigh(noise_cov, info, rank, scalings, projs,
+                                 ch_names)
+    noise_cov.update(eig=eig, eigvec=eigvec)
     return noise_cov
 
 
-def _smart_eigh(C, info, rank, scalings, projs=None, ch_names=None,
-                proj_subspace=False):
+@verbose
+def _smart_eigh(C, info, rank, scalings=None, projs=None,
+                ch_names=None, proj_subspace=False, verbose=None):
     """Compute eigh of C taking into account rank and ch_type scalings."""
+    scalings = _handle_default('scalings_cov_rank', scalings)
     info = info.copy()
     projs = info['projs'] if projs is None else projs
     ch_names = info['ch_names'] if ch_names is None else ch_names
@@ -1422,15 +1401,16 @@ def _smart_eigh(C, info, rank, scalings, projs=None, ch_names=None,
     # Create the projection operator
     proj, ncomp, _ = make_projector(projs, ch_names)
 
+    if isinstance(C, Covariance):
+        C = C['data']
     if ncomp > 0:
         logger.info('    Created an SSP operator (subspace dimension = %d)'
                     % ncomp)
         C = np.dot(proj, np.dot(C, proj.T))
 
-    if rank is None:
-        rank = {}
-    if not isinstance(rank, dict):
-        rank = dict(meg=int(operator.index(rank)))
+    noise_cov = Covariance(C, ch_names, [], projs, 0)
+    rank = compute_rank(noise_cov, rank, scalings, info)
+    assert C.ndim == 2 and C.shape[0] == C.shape[1]
 
     eig = np.zeros(n_chan)
     eigvec = np.zeros((n_chan, n_chan))
@@ -1440,19 +1420,17 @@ def _smart_eigh(C, info, rank, scalings, projs=None, ch_names=None,
         if len(picks) == 0:
             continue
         this_C = C[np.ix_(picks, picks)]
-        this_info = pick_info(_simplify_info(info), picks, copy=False)
-        if rank == 'full':
-            this_rank = len(this_C)
+
+        if ch_type not in rank and ch_type in ('mag', 'grad'):
+            this_rank = rank['meg']  # if there is only one or the other
         else:
-            this_rank = rank.get(ch_type)
-        if this_rank is None:
-            this_rank = _estimate_rank_meeg_cov(this_C, this_info, scalings)
+            this_rank = rank[ch_type]
+
         e, ev, m = _get_ch_whitener(this_C, False, ch_type.upper(), this_rank)
         if proj_subspace:
             # Choose the subspace the same way we do for projections
             e, ev = _eigvec_subspace(e, ev, m)
         eig[picks], eigvec[np.ix_(picks, picks)], mask[picks] = e, ev, m
-
         # XXX : also handle ref for sEEG and ECoG
         if ch_type == 'eeg' and _needs_eeg_average_ref_proj(info) and not \
                 _has_eeg_average_ref_proj(projs):
@@ -1506,17 +1484,12 @@ def regularize(cov, info, mag=0.1, grad=0.1, eeg=0.1, exclude='bads',
         Regularization factor for HBO signals.
     hbr : float (default 0.1)
         Regularization factor for HBR signals.
-    rank : None | int | dict | 'full'
-        Specified rank of the noise covariance matrix. If None (default),
-        the rank is detected automatically.
-        If int, the rank is specified for the MEG
-        channels. A dictionary with entries 'eeg', 'meg' or any other
-        data channel type such as 'seeg' or 'ecog' can be used
-        to specify the rank for each modality. If 'full',
-        the covariance is assumed to be full-rank when regularizing
-        (unless proj=True, in which case projections are accounted for).
+    %(rank_None)s
 
         .. versionadded:: 0.17
+
+        .. versionadded:: 0.18
+           Support for 'info' mode.
     scalings : dict | None
         Data will be rescaled before rank estimation to improve accuracy.
         See :func:`mne.compute_covariance`.
@@ -1541,7 +1514,6 @@ def regularize(cov, info, mag=0.1, grad=0.1, eeg=0.1, exclude='bads',
     if exclude is None:
         raise ValueError('exclude must be a list of strings or "bads"')
 
-    rank, _ = _check_rank_cov(rank, ('',), False)
     if exclude == 'bads':
         exclude = info['bads'] + cov['bads']
 
@@ -1558,6 +1530,8 @@ def regularize(cov, info, mag=0.1, grad=0.1, eeg=0.1, exclude='bads',
         regs['meg'] = mag
     else:
         regs.update(mag=mag, grad=grad)
+    if rank != 'full':
+        rank = compute_rank(cov, rank, scalings, info)
 
     info_ch_names = info['ch_names']
     ch_names_by_type = dict()
@@ -1614,8 +1588,7 @@ def regularize(cov, info, mag=0.1, grad=0.1, eeg=0.1, exclude='bads',
             this_info = pick_info(info, this_picks)
             # Here we could use proj_subspace=True, but this should not matter
             # since this is already in a loop over channel types
-            _, eigvec, mask = _smart_eigh(
-                this_C, this_info, rank, scalings=scalings)
+            _, eigvec, mask = _smart_eigh(this_C, this_info, rank)
             U = eigvec[mask].T
         this_C = np.dot(U.T, np.dot(this_C, U))
 
@@ -1659,7 +1632,7 @@ def _regularized_covariance(data, reg=None, method_params=None, info=None,
     elif not isinstance(reg, str):
         raise ValueError('reg must be a float, str, or None, got %s (%s)'
                          % (reg, type(reg)))
-    method, method_params, rank = _check_method_params(
+    method, method_params = _check_method_params(
         reg, method_params, name='reg', allow_auto=False, rank=rank)
     # use mag instead of eeg here to avoid the cov EEG projection warning
     info = create_info(data.shape[-2], 1000., 'mag') if info is None else info
@@ -1673,11 +1646,9 @@ def _regularized_covariance(data, reg=None, method_params=None, info=None,
     return cov
 
 
-# XXX pca parameter from _get_whitener should be exposed here.
 @verbose
-def compute_whitener(noise_cov, info, picks=None, rank=None,
-                     scalings=None, return_rank=False,
-                     verbose=None):
+def compute_whitener(noise_cov, info, picks=None, rank=None, scalings=None,
+                     return_rank=False, pca=False, verbose=None):
     """Compute whitening matrix.
 
     Parameters
@@ -1687,11 +1658,10 @@ def compute_whitener(noise_cov, info, picks=None, rank=None,
     info : dict
         The measurement info.
     %(picks_good_data_noref)s
-    rank : None | int | dict
-        Specified rank of the noise covariance matrix. If None, the rank is
-        detected automatically. If int, the rank is specified for the MEG
-        channels. A dictionary with entries 'eeg' and/or 'meg' can be used
-        to specify the rank for each modality.
+    %(rank_None)s
+
+        .. versionadded:: 0.18
+           Support for 'info' mode.
     scalings : dict | None
         The rescaling method to be applied. See documentation of
         ``prepare_noise_cov`` for details.
@@ -1699,12 +1669,17 @@ def compute_whitener(noise_cov, info, picks=None, rank=None,
         If True, return the rank used to compute the whitener.
 
         .. versionadded:: 0.15
+    pca : bool
+        If True, return the rank-reduced whitener will be returned.
+
+        .. versionadded:: 0.18
     %(verbose)s
 
     Returns
     -------
-    W : 2d array
-        The whitening matrix.
+    W : ndarray, shape (n_channels, n_channels) or (n_nonzero, n_channels)
+        The whitening matrix, backprojected or not based on whether `pca` is
+        False or True, respectively.
     ch_names : list
         The channel names.
     rank : int
@@ -1722,12 +1697,14 @@ def compute_whitener(noise_cov, info, picks=None, rank=None,
         raise RuntimeError('Not all channels present in noise covariance:\n%s'
                            % missing)
 
-    scalings = _handle_default('scalings_cov_rank', scalings)
     W, _, noise_cov, n_nzero = _get_whitener(
-        noise_cov, info, ch_names, rank, pca=False, scalings=scalings)
+        noise_cov, info, ch_names, rank, pca=pca, scalings=scalings)
 
     # Do the back projection
-    W = np.dot(noise_cov['eigvec'].T, W)
+    if not pca:
+        W = np.dot(noise_cov['eigvec'].T, W)
+    else:
+        assert W.shape[0] == n_nzero
     out = W, ch_names
     if return_rank:
         out += (n_nzero,)
@@ -1748,12 +1725,10 @@ def whiten_evoked(evoked, noise_cov, picks=None, diag=None, rank=None,
     %(picks_good_data)s
     diag : bool (default False)
         If True, whiten using only the diagonal of the covariance.
-    rank : None | int | dict (default None)
-        Specified rank of the noise covariance matrix. If None, the rank is
-        detected automatically. If int, the rank is specified for the MEG
-        channels. A dictionary with entries 'eeg', 'meg' or any other
-        data channel type such as 'seeg' or 'ecog' can be used
-        to specify the rank for each modality.
+    %(rank_None)s
+
+        .. versionadded:: 0.18
+           Support for 'info' mode.
     scalings : dict | None (default None)
         To achieve reliable rank estimation on multiple sensors,
         sensors have to be rescaled. This parameter controls the