+++ mne/utils/__init__.py
@@ -11,8 +11,7 @@ from .check import (check_fname, check_version, check_random_state,
                     _check_event_id, _check_ch_locs, _check_compensation_grade,
                     _check_if_nan, _is_numeric, _ensure_int,  _check_preload,
                     _validate_type, _check_pyface_backend, _check_info_inv,
-                    _check_channels_spatial_filter, _check_rank,
-                    _check_rank_cov)
+                    _check_channels_spatial_filter, _check_rank)
 from .config import (set_config, get_config, get_config_path, set_cache_dir,
                      set_memmap_min_size, get_subjects_dir, _get_stim_channel,
                      sys_info, _get_extra_data_path, _get_root_dir,
@@ -41,6 +40,7 @@ from .numerics import (hashfunc, md5sum, _compute_row_norms,
                        sum_squared, split_list, _gen_events, create_slices,
                        _time_mask, grand_average, object_diff, object_hash,
                        object_size, _apply_scaling_cov, _undo_scaling_cov,
-                       _apply_scaling_array, _undo_scaling_array)
+                       _apply_scaling_array, _undo_scaling_array,
+                       _scaled_array)
 from .mixin import (SizeMixin, GetEpochsMixin, _prepare_read_metadata,
                     _prepare_write_metadata, _FakeNoPandas)